// frontend/src/api/generated/models/accessDto.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * dashboard API
 * OpenAPI spec version: 1.0.0
 */
import type { AccessDtoAccessToken } from './accessDtoAccessToken';
import type { AccessDtoAccessTokenExpiresAt } from './accessDtoAccessTokenExpiresAt';
import type { AccessDtoRefreshToken } from './accessDtoRefreshToken';
import type { AccessDtoRefreshTokenExpiresAt } from './accessDtoRefreshTokenExpiresAt';

export interface AccessDto {
  accessToken?: AccessDtoAccessToken;
  accessTokenExpiresAt?: AccessDtoAccessTokenExpiresAt;
  refreshToken?: AccessDtoRefreshToken;
  refreshTokenExpiresAt?: AccessDtoRefreshTokenExpiresAt;
}